requirements.txt
python-telegram-bot==20.0
openai
youtube_dl
requests
Pillow
